convert graph reference tests to vitest
